+++ src/components/calcite-accordion-item/calcite-accordion-item.tsx
@@ -6,7 +6,7 @@ import {
   h,
   Host,
   Listen,
-  Prop
+  Prop,
 } from "@stencil/core";
 import { UP, DOWN, ENTER, HOME, END, SPACE } from "../../utils/keys";
 import { getElementDir, getElementProp } from "../../utils/dom";
@@ -15,7 +15,7 @@ import { guid } from "../../utils/guid";
 @Component({
   tag: "calcite-accordion-item",
   styleUrl: "calcite-accordion-item.scss",
-  shadow: true
+  shadow: true,
 })
 export class CalciteAccordionItem {
   //--------------------------------------------------------------------------
@@ -59,7 +59,7 @@ export class CalciteAccordionItem {
   componentDidLoad() {
     this.itemPosition = this.getItemPosition();
     this.registerCalciteAccordionItem.emit({
-      position: this.itemPosition
+      position: this.itemPosition,
     });
   }
 
@@ -67,7 +67,10 @@ export class CalciteAccordionItem {
     const dir = getElementDir(this.el);
     return (
       <Host tabindex="0" aria-expanded={this.active.toString()} dir={dir}>
-        <div class="accordion-item-header" onClick={this.itemHeaderClickHander}>
+        <div
+          class="accordion-item-header"
+          onClick={this.itemHeaderClickHandler}
+        >
           <div class="accordion-item-header-text">
             <span class="accordion-item-title">{this.itemTitle}</span>
             <span class="accordion-item-subtitle">{this.itemSubtitle}</span>
@@ -146,7 +149,7 @@ export class CalciteAccordionItem {
   private iconType = getElementProp(this.el, "icon-type", "chevron");
 
   /** handle clicks on item header */
-  private itemHeaderClickHander = () => this.emitRequestedItem();
+  private itemHeaderClickHandler = () => this.emitRequestedItem();
   //--------------------------------------------------------------------------
   //
   //  Private Methods
@@ -174,7 +177,7 @@ export class CalciteAccordionItem {
 
   private emitRequestedItem() {
     this.calciteAccordionItemSelected.emit({
-      requestedAccordionItem: this.accordionItemId
+      requestedAccordionItem: this.accordionItemId,
     });
   }