+++ src/components/calcite-dropdown/calcite-dropdown.tsx
@@ -1,11 +1,4 @@
-import {
-  Component,
-  Element,
-  h,
-  Host,
-  Listen,
-  Prop,
-} from "@stencil/core";
+import { Component, Element, h, Host, Listen, Prop } from "@stencil/core";
 import {
   DOWN,
   END,
@@ -14,15 +7,16 @@ import {
   HOME,
   SPACE,
   TAB,
-  UP
+  UP,
 } from "../../utils/keys";
 
 import { focusElement } from "../../utils/dom";
+import { GroupRegistration } from "../../interfaces/Dropdown";
 
 @Component({
   tag: "calcite-dropdown",
   styleUrl: "calcite-dropdown.scss",
-  shadow: true
+  shadow: true,
 })
 export class CalciteDropdown {
   //--------------------------------------------------------------------------
@@ -46,6 +40,9 @@ export class CalciteDropdown {
     | "start"
     | "center"
     | "end" = "start";
+
+  /** specify the max items to display before showing the scroller, must be greater than 0 **/
+  @Prop() maxItems: number = 0;
 
   /** specify the theme of the dropdown, defaults to light */
   @Prop({ mutable: true, reflect: true }) theme: "light" | "dark";
@@ -84,12 +81,24 @@ export class CalciteDropdown {
       "[slot=dropdown-trigger]"
     ) as HTMLSlotElement;
     if (!this.sorted) {
-      this.items = this.sortItems(this.items);
+      const groups = this.items.sort(
+        (a, b) => a.position - b.position
+      ) as GroupRegistration[];
+
+      this.maxScrollerHeight = this.getMaxScrollerHeight(groups);
+
+      this.items = groups.reduce(
+        (items, group) => [...items, ...group.items],
+        []
+      );
+
       this.sorted = true;
     }
   }
 
   render() {
+    const { maxScrollerHeight } = this;
+
     return (
       <Host>
         <slot
@@ -97,7 +106,13 @@ export class CalciteDropdown {
           aria-haspopup="true"
           aria-expanded={this.active.toString()}
         />
-        <div class="calcite-dropdown-wrapper" role="menu">
+        <div
+          class="calcite-dropdown-wrapper"
+          role="menu"
+          style={{
+            maxHeight: maxScrollerHeight > 0 ? `${maxScrollerHeight}px` : "",
+          }}
+        >
           <slot />
         </div>
       </Host>
@@ -191,14 +206,14 @@ export class CalciteDropdown {
     }
   }
 
-  @Listen("registerCalciteDropdownGroup") registerCalciteDropdownGroup(
-    e: CustomEvent
-  ) {
-    const items = {
-      items: e.detail.items,
-      position: e.detail.position
-    };
-    this.items.push(items);
+  @Listen("registerCalciteDropdownGroup") registerCalciteDropdownGroup({
+    detail: { items, position, titleEl },
+  }: CustomEvent<GroupRegistration>) {
+    this.items.push({
+      items: items,
+      position: position,
+      titleEl,
+    });
   }
 
   //--------------------------------------------------------------------------
@@ -207,20 +222,44 @@ export class CalciteDropdown {
   //
   //--------------------------------------------------------------------------
 
-  /** trigger element */
-  private trigger: HTMLSlotElement;
-
   /** created list of dropdown items */
   private items = [];
 
+  /** specifies the item wrapper height; it is updated when maxItems is > 0  **/
+  private maxScrollerHeight = 0;
+
   /** keep track of whether the groups have been sorted so we don't re-sort */
   private sorted = false;
+
+  /** trigger element */
+  private trigger: HTMLSlotElement;
 
   //--------------------------------------------------------------------------
   //
   //  Private Methods
   //
   //--------------------------------------------------------------------------
+
+  private getMaxScrollerHeight(groups: GroupRegistration[]): number {
+    const { maxItems } = this;
+    let itemsToProcess = 0;
+    let maxScrollerHeight = 0;
+
+    groups.forEach((group) => {
+      if (maxItems > 0 && itemsToProcess < maxItems) {
+        maxScrollerHeight += group?.titleEl?.offsetHeight || 0;
+
+        group.items.forEach((item) => {
+          if (itemsToProcess < maxItems) {
+            maxScrollerHeight += item.offsetHeight;
+            itemsToProcess += 1;
+          }
+        });
+      }
+    });
+
+    return maxScrollerHeight;
+  }
 
   private closeCalciteDropdown() {
     this.active = false;
@@ -229,7 +268,7 @@ export class CalciteDropdown {
 
   private focusOnFirstActiveOrFirstItem(): void {
     this.getFocusableElement(
-      this.items.find(item => item.active) || this.items[0]
+      this.items.find((item) => item.active) || this.items[0]
     );
   }
 
@@ -276,15 +315,10 @@ export class CalciteDropdown {
     const animationDelayInMs = 50;
 
     if (this.active) {
-      setTimeout(() => this.focusOnFirstActiveOrFirstItem(), animationDelayInMs);
+      setTimeout(
+        () => this.focusOnFirstActiveOrFirstItem(),
+        animationDelayInMs
+      );
     }
   }
-
-  private sortItems = (items: any[]): any[] =>
-    items
-      .sort((a, b) => a.position - b.position)
-      .concat.apply(
-        [],
-        this.items.map(item => item.items)
-      );
 }