+++ src/components/calcite-icon/calcite-icon.tsx
@@ -6,19 +6,20 @@ import {
   Host,
   Prop,
   State,
-  Watch
+  Watch,
 } from "@stencil/core";
 import { CSS } from "./resources";
 import { getElementDir } from "../../utils/dom";
 import { fetchIcon, scaleToPx } from "./utils";
 import { Scale } from "../../interfaces/Icon";
 import { Theme } from "../../interfaces/common";
+import { CalciteIconPath, CalciteMultiPathEntry } from "@esri/calcite-ui-icons";
 
 @Component({
   assetsDirs: ["assets"],
   tag: "calcite-icon",
   styleUrl: "calcite-icon.scss",
-  shadow: true
+  shadow: true,
 })
 export class CalciteIcon {
   //--------------------------------------------------------------------------
@@ -37,18 +38,10 @@ export class CalciteIcon {
   //--------------------------------------------------------------------------
 
   /**
-   * When true, the icon will be filled.
-   */
-  @Prop({
-    reflect: true
-  })
-  filled: boolean = false;
-
-  /**
    * The name of the icon to display. The value of this property must match the icon name from https://esri.github.io/calcite-ui-icons/.
    */
   @Prop({
-    reflect: true
+    reflect: true,
   })
   icon: string = null;
 
@@ -56,7 +49,7 @@ export class CalciteIcon {
    * When true, the icon will be mirrored when the element direction is 'rtl'.
    */
   @Prop({
-    reflect: true
+    reflect: true,
   })
   mirrored: boolean = false;
 
@@ -64,7 +57,7 @@ export class CalciteIcon {
    * Icon scale. Can be "s" | "m" | "l".
    */
   @Prop({
-    reflect: true
+    reflect: true,
   })
   scale: Scale = "m";
 
@@ -80,7 +73,7 @@ export class CalciteIcon {
    * Icon theme. Can be "light" or "dark".
    */
   @Prop({
-    reflect: true
+    reflect: true,
   })
   theme: Theme;
 
@@ -113,27 +106,31 @@ export class CalciteIcon {
     const dir = getElementDir(el);
     const size = scaleToPx[scale];
     const semantic = !!textLabel;
-
+    const paths = [].concat(pathData || "");
     return (
       <Host
         aria-label={semantic ? textLabel : null}
         role={semantic ? "img" : null}
       >
-        {pathData ? (
-          <svg
-            class={{
-              [CSS.mirrored]: dir === "rtl" && mirrored,
-              svg: true
-            }}
-            xmlns="http://www.w3.org/2000/svg"
-            fill="currentColor"
-            height={size}
-            width={size}
-            viewBox={`0 0 ${size} ${size}`}
-          >
-            <path d={pathData} />
-          </svg>
-        ) : null}
+        <svg
+          class={{
+            [CSS.mirrored]: dir === "rtl" && mirrored,
+            svg: true,
+          }}
+          xmlns="http://www.w3.org/2000/svg"
+          fill="currentColor"
+          height={size}
+          width={size}
+          viewBox={`0 0 ${size} ${size}`}
+        >
+          {paths.map((path: string | CalciteMultiPathEntry) =>
+            typeof path === "string" ? (
+              <path d={path} />
+            ) : (
+              <path d={path.d} opacity={"opacity" in path ? path.opacity : 1} />
+            )
+          )}
+        </svg>
       </Host>
     );
   }
@@ -147,7 +144,7 @@ export class CalciteIcon {
   private intersectionObserver: IntersectionObserver;
 
   @State()
-  private pathData: string;
+  private pathData: CalciteIconPath;
 
   @State()
   private visible = false;
@@ -159,16 +156,15 @@ export class CalciteIcon {
   //--------------------------------------------------------------------------
 
   @Watch("icon")
-  @Watch("filled")
   @Watch("scale")
   private async loadIconPathData(): Promise<void> {
-    const { filled, icon, scale, visible } = this;
+    const { icon, scale, visible } = this;
 
     if (!Build.isBrowser || !icon || !visible) {
       return;
     }
 
-    this.pathData = await fetchIcon({ icon, scale, filled });
+    this.pathData = await fetchIcon({ icon, scale });
   }
 
   private waitUntilVisible(callback: () => void): void {
@@ -182,8 +178,8 @@ export class CalciteIcon {
     }
 
     this.intersectionObserver = new IntersectionObserver(
-      entries => {
-        entries.forEach(entry => {
+      (entries) => {
+        entries.forEach((entry) => {
           if (entry.isIntersecting) {
             this.intersectionObserver.disconnect();
             this.intersectionObserver = null;